Restructure code, catch illegal corner cases, remove premature compute
