+++ dask/array/core.py
@@ -1577,453 +1577,8 @@ class Array(DaskMethodsMixin):
         return self._scalarfunc(int)
 
     def __setitem__(self, key, value):
-        from .routines import where
-
-        def parse_indices(shape, indices):
-            """Reformat the indices.
-
-             The aim of this is is convert the indices to a
-             standardised form so that it is easier a) to check them
-             for validity, and b) to ascertain which chunks are touched
-             by the indices.
-
-             Note that indices which are decreasing (such as as
-             ``slice(7, 3, -1)`` and ``[6, 2, 1]``) are recast as
-             increasing indices (``slice(4, 8, 1)`` and ``[1, 2, 6]``
-             respectively) to facilitate finding which chunks are
-             touched by the indices. The make sure that the correct
-             values are still assigned, the value is (effectively)
-             reversed along the appropriate dimensions at compute time.
-
-             Parameters
-             ----------
-             shape : sequence of ints
-                 The shape of the global array.
-             indices : tuple
-                 The given indices for assignment.
-
-             Returns
-             -------
-             parsed_indices : list
-                 The reformated indices that are equivalent to the
-                 input indices.
-            indices_shape : list
-                 The shape of the parsed indices. E.g. indices of
-                 ``(slice(0,2), 5, [4,2,1)`` will have shape ``[2,3]``.
-             mirror : list
-                 The dimensions that need to be reversed in the
-                 assigment value, prior to assignment.
-
-             Examples
-             --------
-             >>> parse_indices((8,), (slice(1, -1),))
-             (slice(1, 7, 1)] [6], [])
-             >>> parse_indices((8,), ([1, 2, 4, 6],))
-             (array([1, 2, 4, 6]), [4], [])
-             >>> parse_indices((8,), (slice(-1, 2, -1),))
-             (slice(3, 8, 1), [5], [0])
-             >>> parse_indices((8,), ([6, 4, 2, 1],))
-             (array([1, 2, 4, 6]), [4], [0])
-
-            """
-            parsed_indices = []
-            indices_shape = []
-            mirror = []
-
-            if not isinstance(indices, tuple):
-                indices = (indices,)
-
-            # Initialize the list of parsed indices as the input
-            # indices with any Ellipsis objects expanded
-            length = len(indices)
-            n = len(shape)
-            ndim = n
-            for index in indices:
-                if index is Ellipsis:
-                    m = n - length + 1
-                    parsed_indices.extend([slice(None)] * m)
-                    n -= m
-                else:
-                    parsed_indices.append(index)
-                    n -= 1
-
-                length -= 1
-
-            len_parsed_indices = len(parsed_indices)
-
-            if ndim and len_parsed_indices > ndim:
-                raise IndexError(
-                    f"Invalid indices {parsed_indices} for array with shape {shape}"
-                )
-
-            if len_parsed_indices < ndim:
-                parsed_indices.extend([slice(None)] * (ndim - len_parsed_indices))
-
-            if not ndim and parsed_indices:
-                raise IndexError("too many indices for array")
-
-            n_lists = 0
-
-            for i, (index, size) in enumerate(zip(parsed_indices, shape)):
-                is_slice = isinstance(index, slice)
-                if is_slice:
-                    # Index is a slice
-                    start, stop, step = index.indices(size)
-                    if step < 0 and stop == -1:
-                        stop = None
-
-                    index = slice(start, stop, step)
-
-                elif isinstance(index, (int, np.integer)):
-                    # Index is an integer
-                    index = int(index)
-                else:
-                    n_lists += 1
-                    if n_lists > 1:
-                        raise NotImplementedError(
-                            "dask is currently limited to at most one "
-                            "dimension's assignment index being a "
-                            "1-d array of integers or booleans. "
-                            f"Got: {indices}"
-                        )
-
-                    convert_to_positive = True
-                    if getattr(
-                        getattr(index, "dtype", None), "kind", None
-                    ) == "b" or isinstance(index[0], bool):
-                        # Index is a sequence of booleans, so convert
-                        # the booleans to non-negative integers. We're
-                        # assuming that anything with a dtype
-                        # attribute also has a size attribute.
-                        if len(index) != size:
-                            raise IndexError(
-                                f"Incorrect number (len(index)) of boolean "
-                                f"indices for dimension with size {size}: "
-                                f"{index}"
-                            )
-                        index = where(index)[0].compute_chunk_sizes()
-                        convert_to_positive = False
-
-                    if not np.shape(index):
-                        if index < 0:
-                            index += size
-
-                        index = slice(index, index + 1, 1)
-                        is_slice = True
-                    else:
-                        len_index = len(index)
-                        if len_index == 1:
-                            index = index[0]
-                            if index < 0:
-                                index += size
-
-                            index = int(index)
-                            index = slice(index, index + 1, 1)
-                            is_slice = True
-                        elif len_index:
-                            if convert_to_positive:
-                                # Convert to non-negative integer 1-d
-                                # array
-                                index = asanyarray(index)
-                                index = where(index < 0, index + size, index)
-
-                            steps = index[1:] - index[:-1]
-                            step = steps[0]
-                            if step and not (steps - step).any():
-                                # Replace the 1-d array index with a
-                                # slice
-                                if step > 0:
-                                    start, stop = index[0], index[-1] + 1
-                                elif step < 0:
-                                    start, stop = index[0], index[-1] - 1
-
-                                if stop < 0:
-                                    stop = None
-
-                                index = slice(start, stop, step)
-                                is_slice = True
-                            else:
-                                if (
-                                    (step > 0 and (steps <= 0).any())
-                                    or (step < 0 and (steps >= 0).any())
-                                    or not step
-                                ):
-                                    raise NotImplementedError(
-                                        "1-d integer array assignment indices "
-                                        "are currently limited to strictly "
-                                        "monotonic sequences. "
-                                        f"Got: {indices}"
-                                    )
-
-                                if step < 0:
-                                    # The array is strictly
-                                    # monotonically decreasing, so
-                                    # reverse it so that it's strictly
-                                    # monotonically increasing. Make
-                                    # a note that this dimension will
-                                    # need reversing later.
-                                    index = index[::-1]
-                                    mirror.append(i)
-                                    step = -step
-                        else:
-                            # This is an empty slice
-                            index = slice(0, 0, 1)
-
-                if is_slice and index.step < 0:
-                    # If the slice step is negative, then transform
-                    # the original slice to a new slice with a
-                    # positive step such that the result of the new
-                    # slice is the reverse of the result of the
-                    # original slice.
-                    #
-                    # For example, if the original slice is
-                    # slice(6,0,-2) then the new slice will be
-                    # slice(2,7,2):
-                    #
-                    # >>> a = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9]
-                    # >>> a[slice(6, 0, -2)]
-                    # [6, 4, 2]
-                    # >>> a[slice(2, 7, 2)]
-                    # [2, 4, 6]
-                    # a[slice(6, 0, -2)] == list(reversed(a[slice(2, 7, 2)]))
-                    # True
-                    start, stop, step = index.indices(size)
-                    step *= -1
-                    div, mod = divmod(start - stop - 1, step)
-                    div_step = div * step
-                    start -= div_step
-                    stop = start + div_step + 1
-
-                    index = slice(start, stop, step)
-                    mirror.append(i)
-
-                parsed_indices[i] = index
-
-            # Find the shape of the indices. E.g. indices of
-            # (slice(0,2), 5, [4,2,1) will have shape [2, 3]. Note
-            # that integer indices are not inclded in the shape.
-            for index in parsed_indices:
-                if isinstance(index, slice):
-                    # Index is a slice object
-                    start, stop, step = index.indices(size)
-                    div, mod = divmod(stop - start, step)
-                    if div <= 0:
-                        indices_shape.append(0)
-                    else:
-                        if mod != 0:
-                            div += 1
-
-                        indices_shape.append(div)
-                elif not isinstance(index, int):
-                    # Index is a sequence of integers
-                    indices_shape.append(len(index))
-
-            return parsed_indices, indices_shape, mirror
-
-        def setitem(
-            array,
-            value=None,
-            block_info=None,
-            indices=None,
-            non_broadcast_dimensions=None,
-            offset=None,
-            base_value_indices=None,
-            mirror=None,
-            value_common_shape=None,
-        ):
-            """Function to be mapped across all blocks which assigns new values to elements of each block.
-
-            This function should not be used directly, instead, a new
-            partial function with value set should be used, e.g.::
-
-                self.map_blocks(functools.partial(setitem, value=x), **kwargs)
-
-            Parameters
-            ----------
-            value : array-like
-                The value from which to assign to the indices.
-            block_info:
-            indices : sequence of slice or int
-                A reformated version of the original indices that were
-                passed to __setitem__.
-            non_broadcast_dimensions : list of int
-                The dimensions of value which do not need to be
-                broadcast against the subspace defined by the indices.
-            offset: `int`
-                The difference in the relative positions of a
-                dimension in value and the corresponding dimension in
-                self. A positive value means the dimension position is
-                higher (further to the right) in self then value.
-            base_value_indices : list of slice or None
-                The indices used for initialising the selection from
-                value. slice(None) elements are unchanged, but an
-                element of None is replaced by an appropriate slice.
-            mirror: list of int
-                The dimensions that need to be reversed in the value,
-                prior to assignment.
-            value_common_shape: tuple of int
-                The shape of those dimensions of value which
-                correspond to dimensions of self.
-
-            Returns
-            -------
-            array : numpy.ndarray
-                The array for the block with assigned elements.
-
-            """
-            array_location = block_info[None]["array-location"]
-
-            # See if this block overlaps the indices
-            overlaps = True
-            block_indices = []
-            subset_shape = []
-            preceeding_size = []
-
-            for index, (loc0, loc1), size in zip(
-                indices, array_location, block_info[None]["chunk-shape"]
-            ):
-                integer_index = isinstance(index, int)
-                if isinstance(index, slice):
-                    # Index is a slice
-                    stop = size
-                    if index.stop < loc1:
-                        stop -= loc1 - index.stop
-
-                    start = index.start - loc0
-                    if start < 0:
-                        # Make start positive
-                        start %= index.step
-
-                    if start >= stop:
-                        # This block does not overlap the indices
-                        overlaps = False
-                        break
-
-                    # Still here? Then this block does overlap the
-                    # indices
-                    step = index.step
-                    block_index = slice(start, stop, step)
-                    block_index_size, rem = divmod(stop - start, step)
-                    if rem:
-                        block_index_size += 1
-
-                    # Find how many elements of 'value' precede this
-                    # block along this dimension. Note that it is
-                    # assumed that the slice step is positive, as will
-                    # be the case for reformatted indices.
-                    pre = index.indices(loc0)
-                    n_preceeding, rem = divmod(pre[1] - pre[0], step)
-                    if rem:
-                        n_preceeding += 1
-                elif integer_index:
-                    # Index is an integer
-                    if not loc0 <= index < loc1:
-                        # This block does not overlap the index
-                        overlaps = False
-                        break
-
-                    block_index = index - loc0
-                else:
-                    # Index is a list of integers
-                    block_index = [i - loc0 for i in index if loc0 <= i < loc1]
-                    if not block_index:
-                        # This block does not overlap the indices
-                        overlaps = False
-                        break
-
-                    # Still here? Then this block does overlap the
-                    # indices
-                    block_index_size = len(block_index)
-                    if block_index_size == 1:
-                        block_index = slice(block_index[0], block_index[0] + 1)
-                    else:
-                        index0 = block_index[0]
-                        step = block_index[1] - index0
-                        if step > 0:
-                            start, stop = index0, block_index[-1] + 1
-                        elif step < 0:
-                            start, stop = index0, block_index[-1] - 1
-
-                        if block_index == list(range(start, stop, step)):
-                            # Replace the list of integers with a slice object
-                            if stop < 0:
-                                stop = None
-
-                            block_index = slice(start, stop, step)
-
-                    # Find how many elements of 'value' precede this
-                    # block along this dimension. Note that it is
-                    # assumed that the sequence is strictly
-                    # monotonically increasing, as will be the case
-                    # for reformatted indices.
-                    n_preceeding = sum(1 for i in index if i < loc0)
-
-                block_indices.append(block_index)
-                if not integer_index:
-                    preceeding_size.append(n_preceeding)
-                    subset_shape.append(block_index_size)
-
-            if not overlaps:
-                # This block does not overlap the indices, so return
-                # the block unchanged.
-                return array
-
-            # Still here? Then this block overlaps the indices and so
-            # needs to have some of its elements assigned.
-
-            # Initialise the indices of 'value' that define the parts
-            # of it which are to be assigned to this block
-            value_indices = base_value_indices[:]
-
-            for i in non_broadcast_dimensions:
-                j = i + offset
-                start = preceeding_size[j]
-                value_indices[i] = slice(start, start + subset_shape[j])
-
-            # Reverse the indices to 'value' if required as a
-            # consequence of reformatting the original indices.
-            for i in mirror:
-                size = value_common_shape[i]
-                start, stop, step = value_indices[i].indices(size)
-                size -= 1
-                start = size - start
-                stop = size - stop
-                if stop < 0:
-                    stop = None
-
-                value_indices[i] = slice(start, stop, -1)
-
-            if value.ndim > len(indices):
-                # 'value' has more dimensions than self, so add a
-                # leading Ellipsis to the indices of 'value'.
-                value_indices.insert(0, Ellipsis)
-
-            # Get the part of 'value' that is to be assigned to
-            # elements of this block
-            v = np.asanyarray(value[tuple(value_indices)])
-
-            if not np.ma.isMA(array) and np.ma.isMA(v):
-                # The block is not masked but the assignment is
-                # masking elements, so turn the non-masked block into
-                # a masked one.
-                array = array.view(np.ma.MaskedArray)
-
-            # Assign elements of 'value' to elements of this block
-            if not array.flags["WRITEABLE"]:
-                # array is read-only, so need to copy it first
-                # (e.g. this might be the case if the dask array was
-                # created by da.empty)
-                array = array.copy()
-            elif sys.getrefcount(array) > 2:
-                # Copy the array if it is in use elsewhere, to avoid
-                # corrupting other objects.
-                array = array.copy()
-
-            array[tuple(block_indices)] = v
-
-            return array
-
-        # Start of main __setitem__ code
+        from .routines import where  # , count_nonzero
+        from .slicing import parse_assignment_indices, setitem
 
         # Keep the "where" method for cases when key is an Array
         if isinstance(key, Array):
@@ -2048,7 +1603,7 @@ class Array(DaskMethodsMixin):
         # assignment via map_blocks
 
         # Reformat input indices
-        indices, indices_shape, mirror = parse_indices(self.shape, key)
+        indices, indices_shape, mirror = parse_assignment_indices(key, self.shape)
 
         # Cast 'value' as a dask array
         if value is np.ma.masked:
@@ -2134,7 +1689,11 @@ class Array(DaskMethodsMixin):
             elif a == b and b != 1:
                 base_value_indices.append(None)
                 non_broadcast_dimensions.append(i)
-            else:
+            elif a is None and b != 1:
+                base_value_indices.append(None)
+                non_broadcast_dimensions.append(i)
+            elif a is not None:
+                # Can't check  ...
                 raise ValueError(
                     f"Can't broadcast data with shape {value_common_shape} "
                     f"across shape {tuple(indices_shape)}"
